입고등록 바코드 삭제

+++ newper_admin/src/main/java/com/newper/repository/GoodsRepo.java
@@ -3,5 +3,5 @@ package com.newper.repository;
 import com.newper.entity.Goods;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GoodsRepo extends JpaRepository<Goods, Integer> {
+public interface GoodsRepo extends JpaRepository<Goods, Long> {
 }
